services/users/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.6.1
Flask-Testing==0.7.1
gunicorn==19.9.0
coverage==4.5.3
flake8==3.7.7
flask-debugtoolbar==0.10.1
flask-cors==3.0.7
<<<<<<< HEAD
flask-migrate==2.4.0
flask-bcrypt==0.7.1
=======
flask-migrate==2.4.0
>>>>>>> 56ca3da6228634e582d203ac043c663a48e90ac1
